adding iqcare 2.0 lable

+++ Solutions/IQCare.Core/IQCare.PMTCT.BusinessProcess/Commands/VisitDetailsCommand.cs
@@ -15,7 +15,7 @@ namespace IQCare.PMTCT.BusinessProcess.Commands
         public DateTime Lmp { get; set; }
         public DateTime Edd { get; set; }
         public int Gestation { get; set; }
-        public int ? AgeAtMenarche { get; set; }
+        public Decimal ? AgeAtMenarche { get; set; }
         public int ParityOne { get; set; }
         public int ParityTwo { get; set; }
         public int Gravidae { get; set; }
